gptJoRevisor/fastApiProject1/requirements.txt
# --- FILE: requirements.txt ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch>=2.2.0
tokenizers==0.19.1
sentencepiece
python-dotenv
motor       # driver async do MongoDB
